+++ src/main/java/org/folio/marc4ld/util/Constants.java
@@ -34,6 +34,7 @@ public class Constants {
   public static final String TAG_111 = "111";
   public static final String TAG_245 = "245";
   public static final String TAG_257 = "257";
+  public static final String TAG_545 = "545";
   public static final String TAG_700 = "700";
   public static final String TAG_710 = "710";
   public static final String TAG_711 = "711";