+++ NEWS.md
@@ -1,5 +1,6 @@
 ## 2.0.0 (IN PROGRESS)
 - Update Mapping for 300$b & $e (Physical description & Accompanying material) [MODLD-508](https://folio-org.atlassian.net/browse/MODLD-508)
+- Convert marc field 545 to Resource description [MODLD-337](https://folio-org.atlassian.net/browse/MODLD-337)
 
 ## 1.0.2 (03-12-2025)
 - Initial release